+++ src/constants/SlideTransition.ts
@@ -1,0 +1,11 @@
+export enum SlideTransition {
+  fadeIn = "fadeIn",
+  slideRight = "slideRight",
+  slideLeft = "slideLeft",
+  slideUp = "slideUp",
+  slideDown = "slideDown",
+  zoomIn = "zoomIn",
+  zoomOut = "zoomOut",
+  rotateIn = "rotateIn",
+  rotateOut = "rotateOut",
+}