+++ src/models/index.ts
@@ -3,5 +3,6 @@ export * from "./CommandType";
 export * from "./DemoFileCache";
 export * from "./Demos";
 export * from "./ShellSetting";
+export * from "./SlideMetadata";
 export * from "./Subscription";
 export * from "./Theme";